requirements.txt
beautifulsoup4==4.10.0
pandas==1.3.3
Pillow==8.4.0
requests==2.26.0
streamlit==1.2.0
